+++ src/player.h
@@ -113,11 +113,11 @@ public:
 
 	uint16_t getRandomMount() const;
 	uint16_t getCurrentMount() const;
-	void setCurrentMount(uint8_t mountId);
+	void setCurrentMount(uint16_t mountId);
 	bool isMounted() const { return defaultOutfit.lookMount != 0; }
 	bool toggleMount(bool mount);
-	bool tameMount(uint8_t mountId);
-	bool untameMount(uint8_t mountId);
+	bool tameMount(uint16_t mountId);
+	bool untameMount(uint16_t mountId);
 	bool hasMount(const Mount* mount) const;
 	bool hasMounts() const;
 	void dismount();
